Revert "Implement Cofoundr Matching with Basic Relevance Sorting Algorithm"
