+++ src/app/onboarding/form-components/InterestsAndValuesForm.tsx
@@ -3,6 +3,7 @@
 import React, { useEffect, useMemo } from "react";
 import { useForm } from "react-hook-form";
 import FormInput from "@/components/ui/FormInput";
+import { InterestsAndValuesFormData } from "../types";
 
 
 function InterestsAndValuesForm({
